Remove flat spectrum data card and run option
As it transpires, the CORSIKA run script already allows the user to
specify an energy slope of zero. As such, having a completely new file
dedicated to a different value of the ESLOPE steering parameter is
really a waste and adds an extra needless layer of maintenance if in the
future the parameters change.