Move .job scripts to their own directory
This is just in the interests of cleanliness. I forsee a lot of
submission files in the future, so it'd be good to keep them separate to
the CORSIKA and sim_telarray scripts. Maybe in the future these could be
relocated to their own directory in the repo root directory.