Update a few run statuses
The corsika_flat_spectrum_array_02 was erroneously labelled as having
successfully completed, yet every single task failed.

The changing from CANCELLED to FAILED aims to make statuses more
consistent and meaningful. If a job is COMPLETED, it has done so
successfully. If it has FAILED, then it has finished with errors. If it
is MIXED, then it is an array job with tasks having either COMPLETED or
FAILED.